requirements.txt
kafka-python
numpy
keras
tensorflow
google-cloud-pubsub